third_party/WebKit/LayoutTests/fast/canvas/canvas-context-gc-custom-properties.html

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
function gc() {
    if (typeof GCController !== "undefined")
        GCController.collectAll();
    else {
        var gcRec = function (n) {
            if (n < 1)
                return {};
            var temp = {i: "ab" + i + (i / 100000)};
            temp += "foo";
            gcRec(n-1);
        };
        for (var i = 0; i < 1000; i++)
            gcRec(10);
    }
}

function runTest() {
    canvas = document.createElement("canvas");
    context = canvas.getContext("2d");
    context.customProperty = "value";
    assert_equals(context.customProperty, "value");
    context = null;
    gc();
    context = canvas.getContext("2d");
    assert_equals(context.customProperty, "value");
}

async_test(t => {
        window.onload = function() {
            t.step(runTest);
            t.done();
        }
}, 'Verify that the custom properties on a Canvas 2D rendering context object are retained across GCs.');

</script>
